requirements.txt
amqp==5.1.1
apturl==0.5.2
asgiref==3.6.0
async-timeout==4.0.3
bcrypt==3.2.0
billiard==4.1.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.3.2
celery==5.3.4
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.4
command-not-found==0.3
cryptography==41.0.1
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
diff-match-patch==20230430
distro==1.7.0
distro-info==1.1+ubuntu0.1
dj-config-url==0.1.1
dj-database-url==2.0.0
Django==4.2.1
django-admin-interface==0.26.0
django-allauth==0.55.2
django-colorfield==0.8.0
django-cors-headers==4.3.1
django-crontab==0.7.1
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.4
django-import-export==3.2.0
django-jazzmin==2.6.0
django-restframework==0.0.1
django-storages==1.14.2
django-tinymce==3.6.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
duplicity==0.8.21
et-xmlfile==1.1.0
fasteners==0.14.1
flake8==6.1.0
flower==2.0.1
future==0.18.2
google-api-core==2.13.0
google-auth==2.23.4
google-cloud-core==2.3.3
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
googlemaps==4.6.0
greenlet==3.0.1
gunicorn==20.1.0
gyp==0.1
httplib2==0.20.2
humanize==4.8.0
idna==3.4
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
kombu==5.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupPy==1.14
MarkupSafe==2.0.1
mccabe==0.7.0
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.21.6
oauthlib==3.2.2
odfpy==1.4.1
olefile==0.46
openpyxl==3.1.2
pandas==1.3.5
paramiko==2.9.3
paypalrestsdk==1.13.1
pexpect==4.8.0
Pillow==9.5.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.25.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyaes==1.6.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycodestyle==2.11.1
pycparser==2.21
pycups==2.0.1
pyflakes==3.1.0
PyGObject==3.42.1
PyJWT==2.8.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3.post1
pyxdg==0.27
PyYAML==6.0.1
redis==5.0.1
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.4
stripe==5.4.0
systemd-python==234
tablib==3.5.0
Telethon==1.31.0
termcolor==2.3.0
text-unidecode==1.3
tornado==6.3.3
twilio==7.9.3
typing_extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
udatetime==0.0.17
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.0.2
usb-creator==0.3.7
vine==5.0.0
wadllib==1.3.6
wcwidth==0.2.8
whitenoise==6.2.0
xdg==5
xkit==0.0.0
xlrd==2.0.1
xlwt==1.3.0
zipp==1.0.0
